Fix the majority of compilation warnings
There are still 2 types of warnings after this PR
1.  @nowarn annotation does not suppress any warnings
This type of warnings would show up when compiling with Java8, and not when compiling with
Java11. Since we cannot depricate Java8 now, we need to tolerate them for now.

2. lazy value logMessageFormatVersion in class KafkaConfig is deprecated: 3.0
It is a legitimate high-stake warning that we should constantly be reminded about.