[LI-FIXUP] Fix log truncation metrics
This should be a direct fix-up to the commit
- bb83f42 [LI-HOTFIX] Generate logs to help estimate the amount of data loss during ULE (#99)

For the bytesTruncated metric, under the original implementation,
the size was computed *after* truncation during the full truncation path, which results in always 0.

For the messageTruncated metric, the original implementation calculated it based on start/end offset & target offset.
However, in Kafka, truncation is performed on the unit of a batch.
This [KIP](https://cwiki.apache.org/confluence/display/KAFKA/KIP-279%3A+Fix+log+divergence+between+leader+and+follower+after+fast+leader+fail+over)
also mentions the behavior.
> _Since this is in the middle of the batch, it will truncate all the way to ..._.

In that sense, the original calculation ignores the messages from batch.baseOffset to targetOffset.

This patch fixes the calculation by exposing the batch searched when performing the truncation,
and a unit test for the metric is added.