[LI-HOTFIX] Add metric to sample the age of consumed data
TICKET = N/A
LI_DESCRIPTION =
For SKU tiering design, it's useful to know the age of consumed data. However, if we simply record
the age of all consumed batches, it means the consumed data must be loaded into user space memory,
which defeats the purpose of the kernel TLS work.

Thus, instead of recording the age of all consumed batches, we run a sampling
of the consumers' Fetch requests (with a default ratio of 0.01).

EXIT_CRITERIA = N/A