[LI-FIXUP] Populate the error fields of the LiCombinedControlResponse properly
TICKET = N/A
LI_DESCRIPTION =
This reverts commit a2ac1c2f9eeeaa2819ade05e6dfe37a491c7ebc2 (#408)
The original commit is incorrect and is a backward incompatible schema change on the LiCombinedControlResponse.json.
This PR reverts the original schema change, and addresses the original problem properly:
1. when the LiCombinedControl request version is below 1, the response populates the
LeaderAndIsrPartitionErrors field with the error code. When the version is at or greather than 1, it
populates the LeaderAndIsrTopics field.

2. When the LiCombinedControl request version is below 1, the StopReplicaPartitionErrors field of
the LiCombinedControlResponse should be populated according to the StopReplicaPartitionStates of the
LiCombinedControlRequest. When the version is at or greather than 1, the StopReplicaPartitionErrors
field should be populated according to the StopReplicaTopicStates of the LiCombinedControlRequest.

EXIT_CRITERIA = The same as the LiCombinedControlRequest feature.