[LI-HOTFIX] Reject invalid replica assignment cancellations
TICKET = https://issues.apache.org/jira/browse/KAFKA-14424
LI_DESCRIPTION =
When reassigning replicas, kafka runs a sanity check to ensure all of the target replicas are alive
before allowing the reassignment request to proceed.
However, for a request that cancels an ongoing reassignment, there is no such check.
The result is that if the original replicas are offline, the cancellation may result in partitions
without any leaders. This problem has been observed multiple times in our clusters.

This PR adds the sanity check to ensure all of the original replicas are online before approving the
cancellation request.

EXIT_CRITERIA = When the issue is resolved in Apache kafka and the fix is pulled in.