[LI-HOTFIX] Allow ListOffsetsRequest to be made from Kafka Controller
TICKET = LIKAFKA-49919
LI_DESCRIPTION =
As part of the power outage resiliency work, we want to make sure that when there are multiple corrupted brokers coming up, and there is an offline partition belonging on those brokers, the controller elects the broker with the highest offset.

To achieve this, we must allow Kafka Controller to call ListOffsets on any broker, so that it can compare the results and elect the broker with the highest offset to be the leader.

EXIT_CRITERIA = If the corrupted brokers feature is merged into Kafka, we can remove this hotfix.